Simulate outbreaks

Under development. Do not use. This simulator uses a branching process to determine the infectiousness of individuals over time. It determines the dates of events based on user-provided distributions. Distributions can be provided either as distcrete objects, as functions computing the probability mass functions (PMF), or as vectors of numbers taken as the PMF for values on 0, 1, ..., length(input) - 1

simulate_outbreak(duration = 100, population_size = 100, R_values,
  dist_incubation, dist_infectious_period, dist_reporting = NULL)

Arguments

duration
the number of days to run the simulation for

population_size
the number of susceptible hosts to use in the simulation; defaults to 100

R_values
a vector of values to be used as reproduction number; values will be drawn at random from this vector to determine the expected numbero of secondary cases for each new case

dist_incubation
the distribution of the incubation period, i.e. the time interval between infection and onset of symtpoms

dist_infectious_period
the distribution of the infectious period, i.e. the time interval between the moment a case starts showing symptoms (onset) and the moment they infect new secondary cases

dist_reporting
(optional) the distribution of the reporting delay, i.e. the time interval between symptom onset and the date at which the case is notified; if NULL (default) reporting dates will not be simulated

Examples

## make toy distributions (showing the 3 possible input types)
incubation <- c(0, 1, 1, 1, 1) # numbers = unscaled PMF
infectious_period <- make_disc_gamma(10, 7) # distcrete object
reporting <- function(x) dpois(x, 5) # PMF function
set.seed(1)
x <- simulate_outbreak(R = runif(100, 1, 3), # random values on [1;3]
                       dist_incubation= incubation,
                       dist_infectious_period = infectious_period,
                       dist_reporting = reporting)
dim(x)

#> [1] 77  6

head(x)

#>       id infector date_infection date_onset date_report        R
#> 1 eLGzdD     <NA>              0          3          12 1.954460
#> 2 MQ90i6   eLGzdD              4          5          12 2.209867
#> 3 VCEaZ4   MQ90i6              8         11          18 2.284577
#> 4 bEGjSA   eLGzdD             12         13          17 2.321596
#> 5 k6rcYK   eLGzdD             12         16          19 1.764776
#> 6 msP3Ki   bEGjSA             15         16          19 2.514174

tail(x)

#>        id infector date_infection date_onset date_report        R
#> 72 K0450K   S0jiBe             83         84          89 2.321596
#> 73 7a20Ap   xeKC7z             83         86          91 1.403364
#> 74 41R9ih   uijTrA             85         87          91 1.820168
#> 75 oiHVQb   Q2diLw             86         87          91 2.284577
#> 76 vEGjkv   7a20Ap             86         88          92 2.796779
#> 77 uo5Bds   shTo1z             90         94          98 1.490977

if (require(epicontacts)) {
  plot(x)
}

#> Loading required package: epicontacts